+++ include/Ark/VM/Types.hpp
@@ -1,11 +1,11 @@
 #ifndef ark_vm_types
 #define ark_vm_types
 
-#include <cinttypes>
+#include <cstddef>  // std::size_t
 
 namespace Ark::internal
 {
-    using PageAddr_t = uint16_t;
+    using PageAddr_t = std::size_t;
 }
 
 #endif