+++ include/Ark/VM/VM.hpp
@@ -123,7 +123,7 @@ namespace Ark
         State* m_state;
 
         int m_exitCode;     ///< VM exit code, defaults to 0. Can be changed through `sys:exit`
-        int m_ip;           ///< instruction pointer
+        int m_gip;          ///< global instruction pointer
         std::size_t m_pp;   ///< page pointer
         uint16_t m_sp;      ///< stack pointer
         uint16_t m_fc;      ///< current frames count