+++ src/REPL/Repl.cpp
@@ -65,7 +65,7 @@ namespace Ark
             }
 
             // save a valid ip if execution failed
-            m_old_ip = vm.m_ip;
+            m_old_ip = vm.m_gip;
             if(!tmp_code.str().empty())
             {
                 if (state.doString(tmp_code.str()))
@@ -81,12 +81,12 @@ namespace Ark
                         // save good code
                         code = tmp_code.str();
                         // place ip to end of bytecode intruction (HALT)
-                        -- vm.m_ip;
+                        --vm.m_gip;
                     }
                     else
                     {
                         // reset ip if execution failed
-                        vm.m_ip = m_old_ip;
+                        vm.m_gip = m_old_ip;
                     }
 
                     state.reset();